requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.43.0
python-dotenv==1.0.1
python-multipart==0.0.20
openpyxl==3.1.5
psycopg[binary]
psycopg_pool
stripe
PyJWT
bcrypt
httpx==0.28.1
boto3
pillow
pytest
